backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
Pillow==9.2.0
djoser==2.2.2
urllib3==1.26.7
django-filter==23.5
django-cors-headers==3.13.0
psycopg2-binary==2.9.3